# specs/001-rag-chatbot/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Chatbot API
  description: |
    API for the book-grounded RAG chatbot with personalization and translation features.
    All chat responses are derived strictly from indexed book content.
  version: 1.0.0
  contact:
    name: RAG Chatbot Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourbook.com
    description: Production server
...
paths:
...
  /api/translate/{chapterSlug}:
    get:
      tags: [Translation]
      summary: Get chapter translation
      description: |
        Returns the Urdu translation of a chapter. If not cached, generates on-demand.
        May return 202 Accepted if translation is in progress.
      operationId: getChapterTranslation
      parameters:
        - name: chapterSlug
          in: path
          required: true
          schema:
            type: string
          example: chapter-1-introduction
        - name: language
          in: query
          schema:
            type: string
            default: ur
            enum: [ur]
      responses:
        '200':
          description: Translation available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '202':
          description: Translation in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationPendingResponse'
        '404':
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Translation service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Translation]
      summary: Request chapter translation
      description: Initiates translation of a chapter if not already cached
      operationId: requestChapterTranslation
      parameters:
        - name: chapterSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequest'
      responses:
        '200':
          description: Translation already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '202':
          description: Translation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationPendingResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TranslationRequest:
      type: object
      required: [language]
      properties:
        language:
          type: string
          enum: [ur]
          default: ur
    TranslationResponse:
      type: object
      required: [chapter_slug, language, content, created_at]
      properties:
        chapter_slug:
          type: string
        language:
          type: string
        content:
          type: string
          description: Translated chapter content
        created_at:
          type: string
          format: date-time
    TranslationPendingResponse:
      type: object
      required: [chapter_slug, language, status]
      properties:
        chapter_slug:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [pending, in_progress]
        estimated_seconds:
          type: integer
          description: Estimated time to completion

    # Auth
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - NOT_FOUND
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Additional error details (dev mode only)